Refactor observability profile path
