+++ packages/shared/src/getObservabilityCachePath.ts
@@ -1,0 +1,5 @@
+import { getReplayPath } from "./getReplayPath";
+
+export function getObservabilityCachePath(...path: string[]) {
+  return getReplayPath("observability-profile", ...path);
+}
