--- packages/shared/src/identifier/cachePath.ts
@@ -1,3 +1,0 @@
-import { getReplayPath } from "../getReplayPath";
-
-export const cachePath = getReplayPath("observability-profile", "device.json");
